letsencrypt: Enable alternative config directory
The motivation here is that certbot puts its certificates and keys (etc) in
the path pointed to by config-dir. By default that's in /etc, but this puts
it on a cloud volume so it persists when a VM instance gets replaced.
(Otherwise, replacing a VM would mint all the certs at once, and that would
be the worst possible time to get throttled by the ACME API servers.)