mail: Set up Amazon SES relay for one domain
I'm pretty dissatisfied with how this works with Terraform, since most of
this has to be generated by clicking arounsd in the AWS console and copied
into Terraform after the fact. At least some of that can't be helped,
since there's an out-of-band process to file a ticket with Amazon to each
domain enabled.